Modification of implementation

+++ src/main/java/airline/model/TablesColumns.java
@@ -14,7 +14,7 @@ public class TablesColumns {
 
     private String type;
 
-    private short dataType;
+    private int dataType;
 
     public String getName() {
         return name;
@@ -40,11 +40,11 @@ public class TablesColumns {
         this.table = table;
     }
 
-    public short getDataType() {
+    public int getDataType() {
         return dataType;
     }
 
-    public void setDataType(short dataType) {
+    public void setDataType(int dataType) {
         this.dataType = dataType;
     }
